fix(stylelint+scripts): clear lint baseline so pre-push gate passes
Two small fixes so 'bun run lint' exits 0 (required by the pre-push hook):

1. Quiet typescript-eslint(restrict-template-expressions) in no-unguarded-hover-colors.cjs by coercing the union-typed ruleMessages args with String(). stylelint typings the args as 'string | number | boolean | ... | RegExp', which the type-aware oxlint pass rejects in a template literal.

2. Break the pre-existing import cycle between generate-component-artifacts.ts and generate-component-constraints.ts by extracting discoverComponentDirectories into discover-component-directories.ts. The orchestrator re-exports the symbol so callers that read it from generate-component-artifacts.ts keep working.
