+++ packages/components/scripts/generate-component-artifacts.ts
@@ -20,11 +20,14 @@
  */
 
 import { existsSync } from 'node:fs';
-import { readdir } from 'node:fs/promises';
 import { join } from 'node:path';
 
 import prettier from 'prettier';
 
+import {
+  discoverComponentDirectories,
+  type DiscoveredComponent,
+} from './discover-component-directories.ts';
 import { checkConstraintsDrift, generateAllConstraints } from './generate-component-constraints.ts';
 import {
   checkExamplesDrift,
@@ -35,6 +38,8 @@ import { generateSchemaForComponent } from './generate-component-schema.ts';
 import { generateVariablesForComponent } from './generate-component-variables.ts';
 import { buildManifest, writeManifest } from './generate-manifest.ts';
 import { renderComponentReadme } from './render-component-readme.ts';
+
+export { discoverComponentDirectories, type DiscoveredComponent };
 
 /**
  * Run prettier over the generated content using the repo's prettier config.
@@ -50,51 +55,6 @@ async function formatGenerated(content: string, filepath: string): Promise<strin
   } catch {
     return content;
   }
-}
-
-export interface DiscoveredComponent {
-  /** Absolute path to the component directory. */
-  directory: string;
-  /** Kebab-case component name (the directory basename). */
-  name: string;
-  /** True for `src/components/experimental/<name>/`. */
-  isExperimental: boolean;
-}
-
-const COMPONENTS_ROOT = join(import.meta.dir, '..', 'src', 'components');
-
-export async function discoverComponentDirectories(): Promise<DiscoveredComponent[]> {
-  const results: DiscoveredComponent[] = [];
-
-  for (const entry of await readdir(COMPONENTS_ROOT, { withFileTypes: true })) {
-    if (!entry.isDirectory()) continue;
-    if (entry.name.startsWith('_')) continue;
-
-    if (entry.name === 'experimental') {
-      const experimentalRoot = join(COMPONENTS_ROOT, 'experimental');
-      for (const subEntry of await readdir(experimentalRoot, { withFileTypes: true })) {
-        if (!subEntry.isDirectory()) continue;
-        if (subEntry.name.startsWith('_')) continue;
-        const directory = join(experimentalRoot, subEntry.name);
-        // A migrated component has both `<name>.svelte` AND `<name>.types.ts` —
-        // the types file is the schema generator's input and the marker of
-        // completed migration. Legacy ad-hoc subdirectories (e.g. chat/) lack it.
-        if (!existsSync(join(directory, `${subEntry.name}.svelte`))) continue;
-        if (!existsSync(join(directory, `${subEntry.name}.types.ts`))) continue;
-        results.push({ directory, name: subEntry.name, isExperimental: true });
-      }
-      continue;
-    }
-
-    if (entry.name === 'icons') continue;
-
-    const directory = join(COMPONENTS_ROOT, entry.name);
-    if (!existsSync(join(directory, `${entry.name}.svelte`))) continue;
-    if (!existsSync(join(directory, `${entry.name}.types.ts`))) continue;
-    results.push({ directory, name: entry.name, isExperimental: false });
-  }
-
-  return results.toSorted((a, b) => a.directory.localeCompare(b.directory));
 }
 
 export interface ComponentArtifacts {
